price type 변경

+++ src/main/java/com/lokoko/domain/product/domain/entity/Product.java
@@ -24,6 +24,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 
@@ -46,7 +47,7 @@ public class Product extends BaseEntity {
     private ProductBrand productBrand;
 
     @Column(nullable = false)
-    private long normalPrice;
+    private BigDecimal normalPrice;
 
     @Column(columnDefinition = "TEXT", nullable = false)
     private String productName;
@@ -82,11 +83,11 @@ public class Product extends BaseEntity {
     private ProductCategory productCategory;
 
     @Enumerated(EnumType.STRING)
-    @Column( length = 20)
+    @Column(length = 20)
     private MainCategory mainCategory;
 
     @Enumerated(EnumType.STRING)
-    @Column( length = 20)
+    @Column(length = 20)
     private MiddleCategory middleCategory;
 
     @Enumerated(EnumType.STRING)
