+++ src/main/java/com/lokoko/domain/user/api/dto/request/AdminProductCreateRequest.java
@@ -9,6 +9,7 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 import jakarta.validation.constraints.Size;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 
@@ -28,7 +29,7 @@ public record AdminProductCreateRequest(
         @Schema(requiredMode = REQUIRED, description = "가격", example = "19900")
         @NotNull
         @PositiveOrZero
-        Long normalPrice,
+        BigDecimal normalPrice,
 
         @Schema(description = "용량 (최대 20자)", example = "30ml")
         @Size(max = 20)
